[WIP] Migrating all Sidebar components over to typescript
